move compute-reward-to-go to objectives/value/functional
