+++ torchrl/envs/transforms/transforms.py
@@ -32,8 +32,9 @@ from torchrl.data.tensor_specs import (
 )
 from torchrl.envs.common import EnvBase, make_tensordict
 from torchrl.envs.transforms import functional as F
-from torchrl.envs.transforms.utils import check_finite, compute_reward2go
+from torchrl.envs.transforms.utils import check_finite
 from torchrl.envs.utils import _sort_keys, step_mdp
+from torchrl.objectives.value.functional import compute_reward2go
 
 try:
     from torchvision.transforms.functional import center_crop
